refactor: break out campaign reward icon tooltip component
